Refactor & organise features/assessment folder
