+++ src/core/assessment/__tests__/coordinator.test.node.ts
@@ -1,6 +1,6 @@
 import { getConfig, setConfigCountry } from '@covid/core/Config';
 import { getInitialPatientState } from '@covid/core/patient/PatientState';
-import { AssessmentCoordinator } from '@covid/features/assessment/AssessmentCoordinator';
+import { AssessmentCoordinator } from '@covid/core/assessment/AssessmentCoordinator';
 
 describe('Checks if user needs to back fill any profile questions', () => {
   it('should be always be false for SE hasRaceEthnicityAnswer config', () => {
