+++ src/features/multi-profile/ConsentForOtherScreen.tsx
@@ -12,7 +12,7 @@ import { BrandedButton, ClickableText, ErrorText, HeaderText, RegularText } from
 import { initialErrorState, ApiErrorState } from '@covid/core/api/ApiServiceErrors';
 import UserService from '@covid/core/user/UserService';
 import { PatientInfosRequest } from '@covid/core/user/dto/UserAPIContracts';
-import AssessmentCoordinator from '@covid/features/assessment/AssessmentCoordinator';
+import AssessmentCoordinator from '@covid/core/assessment/AssessmentCoordinator';
 import i18n from '@covid/locale/i18n';
 import { userService, offlineService } from '@covid/Services';
